pyproject.toml
[tool.poetry]
name = "sagesmodel"
version = "0.1.0"
description = ""
authors = ["Bartosz Mikulski <mail@mikulskibartosz.name>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "==2.1.4"
scikit-learn = "^1.3.0"
prophet = "==1.1.5"

[tool.poetry.dev-dependencies]
black = "==23.12.0"
PyYAML = "==6.0.1"
toml = "==0.10.2"
behave = "==1.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
